Isolate generated subtitle workflow from engine facade
generate_subtitles is a small file-generation workflow and can live outside the remaining engine facade while preserving mcp_video.engine.generate_subtitles for public callers. The burn-in path also reuses the shared quality helper instead of hardcoded preset/CRF flags.

Constraint: Keep existing client/server/CLI imports and subtitle output behavior stable while continuing low-conflict engine decomposition.

Rejected: Merge with the existing subtitles burn-in module | generate_subtitles owns SRT generation and optional burn-in, while subtitles() burns an existing subtitle file; keeping modules separate avoids confusing two APIs.

Confidence: high

Scope-risk: narrow

Directive: Keep SRT-generation behavior separate from existing-file subtitle burn-in unless the public API is redesigned deliberately.

Tested: ruff check mcp_video/engine.py mcp_video/engine_subtitle_generate.py; ruff format --check mcp_video/engine.py mcp_video/engine_subtitle_generate.py; python3 -m pytest tests/test_engine_advanced.py::TestGenerateSubtitles -q --tb=short; python3 -m pytest tests/test_cli.py -k 'generate_subtitles' tests/test_red_team.py -k 'generate_subtitles' -q --tb=short; python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short

Not-tested: Full slow/real-media suite.
